further consolidate apache dependencies
