+++ kifi-backend/modules/scraper/app/com/keepit/scraper/DevScraperProcessorModule.scala
@@ -6,7 +6,8 @@ import com.keepit.scraper.extractor.{ ExtractorFactoryImpl, ExtractorFactory }
 import akka.actor.ActorSystem
 import com.keepit.common.healthcheck.AirbrakeNotifier
 import com.keepit.common.plugin.SchedulingProperties
-import com.keepit.scraper.fetcher.{ ApacheHttpFetcher, HttpFetcher }
+import com.keepit.scraper.fetcher.HttpFetcher
+import com.keepit.scraper.fetcher.apache.ApacheHttpFetcher
 
 case class DevScraperProcessorModule() extends ScrapeProcessorModule {
