+++ kifi-backend/modules/scraper/test/com/keepit/scraper/TestScraperProcessorModule.scala
@@ -6,7 +6,8 @@ import com.google.inject.{ Provider, Provides, Singleton }
 import akka.actor.ActorSystem
 import com.keepit.common.healthcheck.AirbrakeNotifier
 import com.keepit.common.plugin.SchedulingProperties
-import com.keepit.scraper.fetcher.{ApacheHttpFetcher, HttpFetcher}
+import com.keepit.scraper.fetcher.HttpFetcher
+import com.keepit.scraper.fetcher.apache.ApacheHttpFetcher
 
 case class TestScraperProcessorModule() extends ScrapeProcessorModule {
